rename jwt-utils and add readme
